src/azure_ml_training/requirements.txt
torch>=2.0.0
torchvision
pandas
numpy
scikit-learn
mlflow
joblib
azure-ai-ml
azureml-mlflow
